+++ src/GazeServiceProvider.php
@@ -16,6 +16,12 @@ use Naoray\GazeLaravel\Console\CanaryCommand;
 use Naoray\GazeLaravel\Console\CheckCommand;
 use Naoray\GazeLaravel\Console\DoctorCommand;
 use Naoray\GazeLaravel\Console\InstallNerCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyLogsCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyRestartCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyServeCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyStartCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyStatusCommand;
+use Naoray\GazeLaravel\Console\Proxy\ProxyStopCommand;
 use Naoray\GazeLaravel\Install\BinaryInstaller;
 use Naoray\GazeLaravel\Install\LaravelNerFetcher;
 use Naoray\GazeLaravel\Install\NerFetcher;
@@ -166,6 +172,12 @@ class GazeServiceProvider extends ServiceProvider implements DeferrableProvider
                 CanaryCommand::class,
                 BenchCommand::class,
                 InstallNerCommand::class,
+                ProxyServeCommand::class,
+                ProxyStartCommand::class,
+                ProxyStopCommand::class,
+                ProxyRestartCommand::class,
+                ProxyStatusCommand::class,
+                ProxyLogsCommand::class,
             ]);
         }
     }